More informative error handling if wrong channel
Current error message if an incorrect channel is entered is     
self.start_time = float(response['ts'])
KeyError: 'ts'

which is quite unhelpful.  This commit changes this to "Exception: channel_not_found".  Note that the exception contents will be whatever the error message content is, so this generalizes nicely to other errors too.